ft_itoa.c updated

+++ src/libft.h
@@ -58,7 +58,7 @@ char *ft_substr(char const *s, unsigned int start, size_t len);
 char *ft_strjoin(char const *s1, char const *s2);
 char *ft_strtrim(char const *s);
 char **ft_split(char const *s, char c);
-char *ft_itoa(int n);
+char *ft_itoa(long long int n);
 void ft_putchar(char c);
 void ft_putstr(char const *s);
 void ft_putendl(char const *s);
